don’t move a failed build

+++ util/lib/golang_bin.sh
@@ -5,6 +5,10 @@ mkdir -p "${BUILD}/bin" "${BUILD}/lib";
 cd "${MODULEDIR}";
 
 go build ./...;
+
+if [ ! -f "${OUTNAME}" ]; then
+	exit 2
+fi
 mv "${OUTNAME}" "${BUILD}/bin";
 
 for file in ${COPYLIB}; do
